+++ src/main/java/calculator/common/di/IntegerCalculatorDependencyRegistry.java
@@ -6,7 +6,9 @@ import calculator.model.custom_delimiter.service.CustomDelimiterService;
 import calculator.model.custom_delimiter.validator.DefaultCustomDelimiterValidator;
 import calculator.model.custom_delimiter.validator.DelimiterValidator;
 import calculator.model.delimiter.factory.DefaultDelimiterFactory;
+import calculator.model.delimiter.factory.DefaultDelimitersFactory;
 import calculator.model.delimiter.factory.DelimiterFactory;
+import calculator.model.delimiter.factory.DelimitersFactory;
 import calculator.model.delimiter.service.DelimiterService;
 import calculator.model.delimiter.service.IntegerDelimiterService;
 import calculator.model.number.service.NumberService;
@@ -17,9 +19,10 @@ public class IntegerCalculatorDependencyRegistry implements NumberCalculatorDepe
 
     private static final CustomDelimiterPatternMatcher customDelimiterPatternMatcher = new DefaultCustomDelimiterPatternMatcher();
     private static final DelimiterValidator delimiterValidator = new DefaultCustomDelimiterValidator();
+    private static final DelimitersFactory delimitersFactory = new DefaultDelimitersFactory();
     private static final DelimiterFactory delimiterFactory = new DefaultDelimiterFactory(delimiterValidator);
     private static final CustomDelimiterService customDelimiterService = new CustomDelimiterService(delimiterFactory, customDelimiterPatternMatcher);
-    private static final DelimiterService delimiterService = new IntegerDelimiterService(delimiterFactory, customDelimiterService);
+    private static final DelimiterService delimiterService = new IntegerDelimiterService(delimitersFactory, customDelimiterService);
     private static final NumberService numberService = new NumberService();
 
     private static final InputView inputView = new InputView();
