+++ src/main/java/calculator/model/delimiter/factory/DelimiterFactory.java
@@ -2,7 +2,6 @@ package calculator.model.delimiter.factory;
 
 import calculator.model.custom_delimiter.validator.DelimiterValidator;
 import calculator.model.delimiter.domain.Delimiter;
-import calculator.model.delimiter.domain.Delimiters;
 
 public abstract class DelimiterFactory {
 
@@ -11,8 +10,6 @@ public abstract class DelimiterFactory {
     public DelimiterFactory(DelimiterValidator delimiterValidator) {
         this.delimiterValidator = delimiterValidator;
     }
-
-    public abstract Delimiters createDelimiters(Delimiter... additionalDelimiters);
 
     public abstract Delimiter createDelimiter(String value);
 }
