+++ src/main/java/calculator/model/delimiter/factory/DelimitersFactory.java
@@ -1,0 +1,9 @@
+package calculator.model.delimiter.factory;
+
+import calculator.model.delimiter.domain.Delimiter;
+import calculator.model.delimiter.domain.Delimiters;
+
+public abstract class DelimitersFactory {
+
+    public abstract Delimiters createDelimiters(Delimiter... additionalDelimiters);
+}
