+++ src/main/java/calculator/model/delimiter/service/NumberDelimiterService.java
@@ -1,14 +1,14 @@
 package calculator.model.delimiter.service;
 
 import calculator.model.custom_delimiter.service.CustomDelimiterService;
-import calculator.model.delimiter.factory.DelimiterFactory;
+import calculator.model.delimiter.factory.DelimitersFactory;
 
 public abstract class NumberDelimiterService extends DelimiterService {
 
     public NumberDelimiterService(
-            DelimiterFactory delimiterFactory,
+            DelimitersFactory delimitersFactory,
             CustomDelimiterService customDelimiterService) {
 
-        super(delimiterFactory, customDelimiterService);
+        super(delimitersFactory, customDelimiterService);
     }
 }
